requirements.txt
flask==2.0.1
werkzeug==2.0.2
requests==2.26.0
beautifulsoup4==4.10.0
flask-wtf==1.0.0
python-dotenv==0.19.0
gunicorn==20.1.0
pandas==1.3.3
tqdm==4.62.3